Change all .cpp suffix to .cxx suffix
follow the coding standard requirement
