+++ examples/nxflat/tests/hello++/hello++1.cxx
@@ -1,7 +1,7 @@
 /////////////////////////////////////////////////////////////////////////////
-// examples/elf/tests/helloxx/hello++1.c
+// apps/examples/nxflat/tests/hello++/hello++1.cxx
 //
-//   Copyright (C) 2012 Gregory Nutt. All rights reserved.
+//   Copyright (C) 2009 Gregory Nutt. All rights reserved.
 //   Author: Gregory Nutt <gnutt@nuttx.org>
 //
 // Redistribution and use in source and binary forms, with or without
