+++ testing/irtest/enum.cxx
@@ -1,5 +1,5 @@
 /****************************************************************************
- * testing/irtest/enum.cpp
+ * apps/testing/irtest/enum.cxx
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
