+++ testing/irtest/main.cxx
@@ -1,5 +1,5 @@
 /****************************************************************************
- * testing/irtest/main.cpp
+ * apps/testing/irtest/main.cxx
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
